Updated required CLI args: time-path, time-format, and time-relative-start. This way we dont guess what the time column is and what type of data it is.
